+++ commonmark-ext-gfm-tables/src/test/java/org/commonmark/ext/gfm/tables/TablesTest.java
@@ -2,7 +2,9 @@ package org.commonmark.ext.gfm.tables;
 
 import org.commonmark.Extension;
 import org.commonmark.node.Node;
+import org.commonmark.node.SourceSpan;
 import org.commonmark.parser.Parser;
+import org.commonmark.parser.IncludeSourceSpans;
 import org.commonmark.renderer.html.AttributeProvider;
 import org.commonmark.renderer.html.AttributeProviderContext;
 import org.commonmark.renderer.html.AttributeProviderFactory;
@@ -10,11 +12,13 @@ import org.commonmark.renderer.html.HtmlRenderer;
 import org.commonmark.testutil.RenderingTestCase;
 import org.junit.Test;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Map;
 import java.util.Set;
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 public class TablesTest extends RenderingTestCase {
@@ -641,6 +645,54 @@ public class TablesTest extends RenderingTestCase {
                 "</tr>\n" +
                 "</tbody>\n" +
                 "</table>\n"));
+    }
+
+    @Test
+    public void sourceSpans() {
+        Parser parser = Parser.builder()
+                .extensions(EXTENSIONS)
+                .includeSourceSpans(IncludeSourceSpans.BLOCKS)
+                .build();
+        Node document = parser.parse("Abc|Def\n---|---\n|1|2\n 3|four|\n|||\n");
+
+        TableBlock block = (TableBlock) document.getFirstChild();
+        assertEquals(Arrays.asList(SourceSpan.of(0, 0, 7), SourceSpan.of(1, 0, 7),
+                SourceSpan.of(2, 0, 4), SourceSpan.of(3, 0, 8), SourceSpan.of(4, 0, 3)),
+                block.getSourceSpans());
+
+        TableHead head = (TableHead) block.getFirstChild();
+        assertEquals(Arrays.asList(SourceSpan.of(0, 0, 7)), head.getSourceSpans());
+
+        TableRow headRow = (TableRow) head.getFirstChild();
+        assertEquals(Arrays.asList(SourceSpan.of(0, 0, 7)), headRow.getSourceSpans());
+        TableCell headRowCell1 = (TableCell) headRow.getFirstChild();
+        TableCell headRowCell2 = (TableCell) headRow.getLastChild();
+        assertEquals(Arrays.asList(SourceSpan.of(0, 0, 3)), headRowCell1.getSourceSpans());
+        assertEquals(Arrays.asList(SourceSpan.of(0, 4, 3)), headRowCell2.getSourceSpans());
+
+        TableBody body = (TableBody) block.getLastChild();
+        assertEquals(Arrays.asList(SourceSpan.of(2, 0, 4), SourceSpan.of(3, 0, 8), SourceSpan.of(4, 0, 3)), body.getSourceSpans());
+
+        TableRow bodyRow1 = (TableRow) body.getFirstChild();
+        assertEquals(Arrays.asList(SourceSpan.of(2, 0, 4)), bodyRow1.getSourceSpans());
+        TableCell bodyRow1Cell1 = (TableCell) bodyRow1.getFirstChild();
+        TableCell bodyRow1Cell2 = (TableCell) bodyRow1.getLastChild();
+        assertEquals(Arrays.asList(SourceSpan.of(2, 1, 1)), bodyRow1Cell1.getSourceSpans());
+        assertEquals(Arrays.asList(SourceSpan.of(2, 3, 1)), bodyRow1Cell2.getSourceSpans());
+
+        TableRow bodyRow2 = (TableRow) body.getFirstChild().getNext();
+        assertEquals(Arrays.asList(SourceSpan.of(3, 0, 8)), bodyRow2.getSourceSpans());
+        TableCell bodyRow2Cell1 = (TableCell) bodyRow2.getFirstChild();
+        TableCell bodyRow2Cell2 = (TableCell) bodyRow2.getLastChild();
+        assertEquals(Arrays.asList(SourceSpan.of(3, 0, 2)), bodyRow2Cell1.getSourceSpans());
+        assertEquals(Arrays.asList(SourceSpan.of(3, 3, 4)), bodyRow2Cell2.getSourceSpans());
+
+        TableRow bodyRow3 = (TableRow) body.getLastChild();
+        assertEquals(Arrays.asList(SourceSpan.of(4, 0, 3)), bodyRow3.getSourceSpans());
+        TableCell bodyRow3Cell1 = (TableCell) bodyRow3.getFirstChild();
+        TableCell bodyRow3Cell2 = (TableCell) bodyRow3.getLastChild();
+        assertEquals(Collections.emptyList(), bodyRow3Cell1.getSourceSpans());
+        assertEquals(Collections.emptyList(), bodyRow3Cell2.getSourceSpans());
     }
 
     @Override