+++ commonmark/src/main/java/org/commonmark/internal/SourceSpans.java
@@ -1,0 +1,13 @@
+package org.commonmark.internal;
+
+import org.commonmark.node.SourceSpan;
+import org.commonmark.parser.block.ParserState;
+
+public class SourceSpans {
+
+    public static SourceSpan fromState(ParserState parserState, int startIndex) {
+//        int length = parserState.getLine().length();
+//        return SourceSpan.of(parserState.getLineIndex(), startIndex, length - startIndex);
+        return null;
+    }
+}