+++ commonmark/src/main/java/org/commonmark/node/Node.java
@@ -1,5 +1,14 @@
 package org.commonmark.node;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+/**
+ * The base class of all CommonMark AST nodes ({@link Block} and inlines).
+ * <p>
+ * A node can have multiple children, and a parent (except for the root node).
+ */
 public abstract class Node {
 
     private Node parent = null;
@@ -7,6 +16,7 @@ public abstract class Node {
     private Node lastChild = null;
     private Node prev = null;
     private Node next = null;
+    private List<SourceSpan> sourceSpans = null;
 
     public abstract void accept(Visitor visitor);
 
@@ -102,6 +112,35 @@ public abstract class Node {
         if (sibling.prev == null) {
             sibling.parent.firstChild = sibling;
         }
+    }
+
+
+    /**
+     * @return the source spans of this node if included by the parser, an empty list otherwise
+     */
+    public List<SourceSpan> getSourceSpans() {
+        return sourceSpans != null ? Collections.unmodifiableList(sourceSpans) : Collections.<SourceSpan>emptyList();
+    }
+
+    /**
+     * Replace the current source spans with the provided list.
+     *
+     * @param sourceSpans the new source spans to set
+     */
+    public void setSourceSpans(List<SourceSpan> sourceSpans) {
+        this.sourceSpans = new ArrayList<>(sourceSpans);
+    }
+
+    /**
+     * Add a source span to the end of the list.
+     *
+     * @param sourceSpan the source span to add
+     */
+    public void addSourceSpan(SourceSpan sourceSpan) {
+        if (sourceSpans == null) {
+            this.sourceSpans = new ArrayList<>();
+        }
+        this.sourceSpans.add(sourceSpan);
     }
 
     @Override