+++ commonmark/src/main/java/org/commonmark/parser/block/AbstractBlockParser.java
@@ -1,6 +1,7 @@
 package org.commonmark.parser.block;
 
 import org.commonmark.node.Block;
+import org.commonmark.node.SourceSpan;
 import org.commonmark.parser.InlineParser;
 
 public abstract class AbstractBlockParser implements BlockParser {
@@ -22,6 +23,11 @@ public abstract class AbstractBlockParser implements BlockParser {
 
     @Override
     public void addLine(CharSequence line) {
+    }
+
+    @Override
+    public void addSourceSpan(SourceSpan sourceSpan) {
+        getBlock().addSourceSpan(sourceSpan);
     }
 
     @Override